pyproject.toml
[project]
name = "supercharge-ai"
dynamic = ["version"]
description = "Supercharge AI Dev - AI-Dev-Kit + Databricks Asset Bundles integration"
requires-python = ">=3.12, <3.13"
dependencies = [
    "pydantic==2.11.7",
    "databricks-sdk==0.85.0",
    "loguru==0.7.3",
    "python-dotenv==1.1.1",
    "pyyaml==6.0.2",
    "databricks-vectorsearch==0.63",
    "mlflow==3.8.1",
    "databricks-agents==1.8.2",
]

[project.optional-dependencies]
dev = [
    "databricks-connect==17.0.0",
    "ipykernel==6.29.5",
    "pip==25.0.1",
    "pre-commit==4.1.0",
    "pytest==8.3.4",
    "pytest-asyncio==0.24.0",
    "pytest-cov==6.0.0",
    "ruff==0.15.1",
    "mypy==1.8.0",
    "types-PyYAML==6.0.12",
    "bandit==1.7.5",
    "interrogate==1.5.0",
]

ci = [
    "pre-commit==4.1.0",
    "pytest==8.3.4",
    "pytest-asyncio==0.24.0",
    "pytest-cov==6.0.0",
    "ruff==0.15.1",
    "mypy==1.8.0",
    "types-PyYAML==6.0.12",
    "pip-audit==2.6.3",
]

[build-system]
requires = ["setuptools==72.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["supercharge_ai*"]

[tool.setuptools.dynamic]
version = { file = "version.txt" }

[tool.pytest.ini_options]
addopts = "-s --no-header --no-summary"
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 90

[tool.ruff.lint]
select = [
    "F",    # pyflakes
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "B",    # flake8-bugbear
    "I",    # isort
    "UP",   # pyupgrade
    "SIM",  # flake8-simplify
    "ANN",  # flake8-annotations
]

ignore = [
    "ANN204",  # Missing return type annotation for __init__
]

[tool.ruff.lint.per-file-ignores]
# Databricks notebooks: ignore long lines, top-of-file imports-not-first (cell structure),
# and undefined names for runtime-injected globals (dbutils, spark, dbruntime, display)
"notebooks/*.py" = ["E501", "E402", "F821"]
"tests/**/*.py" = ["ANN"]

[tool.ruff.format]
indent-style = "space"
